requirements.txt
Flask==3.1.1
mutagen==1.47.0
Pillow==11.3.0
pywebview==5.4
